LayoutTests/inspector/debugger/reveal-execution-line-expected.txt

Tests that execution line is revealed and highlighted when debugger is paused.

Bug 80306
Debugger was enabled.

Running: testRevealAndHighlightExecutionLine
Set timer for test function.
Execution line revealed and highlighted.
Debugger was disabled.
